+++ src/lollms_client/__init__.py
@@ -9,7 +9,7 @@ from lollms_client.lollms_llm_binding import LollmsLLMBindingManager
 # Import new bindings utils
 from lollms_client.lollms_bindings_utils import list_bindings, get_binding_desc
 
-__version__ = "1.10.0" # Updated version
+__version__ = "1.10.1" # Updated version
 
 # Optionally, you could define __all__ if you want to be explicit about exports
 __all__ = [